requirements.txt
Flask
Flask-Security-Too
Flask-Migrate
Flask-SQLAlchemy
Flask-MongoEngine
SQLAlchemy
mongoengine
pymysql
pymongo
requests
GitPython